ipc: fix mp channel closure to prevent message loss
This commit addresses an issue related to the cleanup of the
multiprocess channel. Previously, when closing the channel, there was a
risk of losing trailing messages. This issue was particularly noticeable
when broadcast message from primary to secondary processes was sent
while a secondary process was closing it's mp channel. In this fix, we
delete mp socket file before stopping mp receive threa.

Signed-off-by: Artemy Kovalyov <artemyko@nvidia.com>